requirements.txt
certifi==2020.12.5
chardet==4.0.0
h11==0.12.0
httpcore==0.12.3
httpx==0.17.1
idna==2.10
joplin-api==1.6.0
oauthlib==3.1.0
py-trello==0.17.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
urllib3==1.26.4